+++ randovania/games/prime1/exporter/hint_namer.py
@@ -18,7 +18,7 @@ class PrimeHintNamer(PrimeFamilyHintNamer):
 
     @override
     @property
-    def color_player(self) -> str:
+    def color_world(self) -> str:
         return "#d4cc33"
 
     @override