+++ randovania/games/prime2/exporter/hint_namer.py
@@ -47,7 +47,7 @@ class EchoesHintNamer(PrimeFamilyHintNamer):
 
     @override
     @property
-    def color_player(self) -> str:
+    def color_world(self) -> str:
         return "#d4cc33"
 
     @override