+++ randovania/generator/filler/filler_configuration.py
@@ -32,6 +32,7 @@ class PlayerPool:
     configuration: BaseConfiguration
     patches: GamePatches
     pickups: list[PickupEntry]
+    all_pickups: list[PickupEntry]
 
 
 @dataclasses.dataclass(frozen=True)